+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
-import Books from './redux/books/Books';
-import Categories from './redux/categories/Categories';
+import Books from './components/Books';
+import Categories from './components/Categories';
 import NavBar from './components/NavBar';
 
 function App() {
