+++ src/components/Categories.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Categories = () => (
+  <div>
+    <h3>Under construction</h3>
+  </div>
+);
+
+export default Categories;
