--- src/redux/categories/Categories.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const Categories = () => (
-  <div>
-    <h3>Under construction</h3>
-  </div>
-);
-
-export default Categories;
